fibonaci2.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 int i;
-int fib[50];
+long long fib[50];
 fib[0] = 1;
 fib[1] = 2;
 for (i = 2; i < 50; i++)
@@ -16,8 +16,8 @@ fib[i] = fib[i - 1] + fib[i - 2];
 }
 for (i = 0; i < 49; i++)
 {
-printf("%d, ", fib[i]);
+printf("%lld, ", fib[i]);
 }
-printf("%d\n", fib[49]);
+printf("%lld\n", fib[49]);
 return (0);
 }
